backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
pillow==10.2.0
